+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 21.11.2018
+
+* [VideoSearch] Started building a RecyclerView adapter
+
 #### 18.11.2018
 
 * [General] Replaced the search button with keyboard search action